+++ functions/api/form/contact-urgent.js
@@ -1,21 +1,25 @@
-import { handleForm } from "../../functions-src/forms.js"
+import { handleForm } from "../../../functions-src/forms.js"
 
 export async function onRequest(context) {
   if (context.request.method !== "POST") {
     return new Response("Invalid request method", { status: 405 });
   }
 
+  const headers = Object.fromEntries(context.request.headers.entries());
   const url = new URL(context.request.url)
+  const usingJson = (headers["content-type"] === "application/json");
 
   const success = await handleForm({
     context,
-    formId: "CONTACT",
+    formId: "URGENT-CONTACT",
     honeypotField: "name",
   });
 
-  if (!success) {
+  if (!success && !usingJson) {
     return new Response("Oops! Something went wrong. Please try submitting the form again.", { status: 500 });
   }
 
-  return Response.redirect(`https://${url.hostname}/contact-success`, 303);
+  return usingJson ? 
+    Response.json({ success }) :
+    Response.redirect(`https://${url.hostname}/contact-success`, 303);
 }
